// frontend-v3/src/app/core/api/models/adicional-input.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { AdicionalOpcaoInput } from '../models/adicional-opcao-input';

/**
 * Dados de entrada para criacao/atualizacao de adicionais
 */
export interface AdicionalInput {

  /**
   * ID da categoria a que o adicional pertence
   */
  id_categoria: number;

  /**
   * Limite maximo (M ou Q)
   */
  limite?: number;

  /**
   * Quantidade minima (apenas para selecao Q)
   */
  minimo?: number;

  /**
   * Nome do grupo de adicionais
   */
  nome: string;

  /**
   * Opcoes de adicionais (itens)
   */
  opcoes: Array<AdicionalOpcaoInput>;

  /**
   * Tipo de selecao: U (unico), M (multiplo), Q (quantidade multipla)
   */
  selecao: string;

  /**
   * Status do adicional (ativo)
   */
  status?: boolean;
}